Fixed and moved all tests
